requirements.txt
# AI SDKs
google-genai
openai

# Data & Utilities
pandas
pydantic
python-dotenv
Faker

# Analysis & Visualization
scipy
seaborn
matplotlib
jupyter
